refactor(all): 패키지 구조를 개선하여 코드 구조화
