+++ src/main/java/christmas/controller/SettlementSystem.java
@@ -1,6 +1,8 @@
 package christmas.controller;
 
-import christmas.domain.*;
+import christmas.domain.entity.Day;
+import christmas.domain.entity.DiscountRecord;
+import christmas.domain.entity.Order;
 import christmas.domain.manager.BonusEventManager;
 import christmas.domain.manager.DiscountManager;
