+++ src/main/java/christmas/domain/entity/Day.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.entity;
 
 import static christmas.system.Constant.CHRISTMAS;
