+++ src/main/java/christmas/domain/entity/MenuItem.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.entity;
 
 public class MenuItem {
