+++ src/main/java/christmas/domain/entity/MenuType.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.entity;
 
 public enum MenuType {
     APPETIZER,
