+++ src/main/java/christmas/domain/manager/BonusEventManager.java
@@ -1,7 +1,7 @@
 package christmas.domain.manager;
 
-import christmas.domain.BonusMenu;
-import christmas.domain.Order;
+import christmas.domain.entity.BonusMenu;
+import christmas.domain.entity.Order;
 
 
 import static christmas.system.Constant.BONUS_MINIMUM_THRESHOLD;
