+++ src/test/java/christmas/domain/entity/BonusItemTest.java
@@ -1,5 +1,6 @@
-package christmas.domain;
+package christmas.domain.entity;
 
+import christmas.domain.entity.BonusItem;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
