+++ src/test/java/christmas/domain/entity/EventBadgeTest.java
@@ -1,5 +1,6 @@
-package christmas.domain;
+package christmas.domain.entity;
 
+import christmas.domain.entity.EventBadge;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
