+++ src/test/java/christmas/domain/entity/MenuItemTest.java
@@ -1,5 +1,7 @@
-package christmas.domain;
+package christmas.domain.entity;
 
+import christmas.domain.entity.MenuItem;
+import christmas.domain.entity.MenuType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
