+++ src/test/java/christmas/domain/entity/OrderItemTest.java
@@ -1,5 +1,8 @@
-package christmas.domain;
+package christmas.domain.entity;
 
+import christmas.domain.entity.MenuItem;
+import christmas.domain.entity.MenuType;
+import christmas.domain.entity.OrderItem;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
