+++ src/test/java/christmas/domain/verifier/subclass/OrderMenuVerifierSubTest.java
@@ -1,6 +1,6 @@
 package christmas.domain.verifier.subclass;
 
-import christmas.domain.Menu;
+import christmas.domain.entity.Menu;
 
 import java.util.HashSet;
 import java.util.Set;
